Move concrete endpoint resolver types to `runtime-api`
This commit addresses https://github.com/awslabs/smithy-rs/pull/2577#discussion_r1169370073
