+++ rust-runtime/aws-smithy-runtime-api/src/client.rs
@@ -21,8 +21,5 @@ pub mod retries;
 /// Runtime plugin type definitions.
 pub mod runtime_plugin;
 
-/// Smithy endpoint resolution runtime plugins
-pub mod endpoints;
-
 /// Smithy auth runtime plugins
 pub mod auth;
